refactor: `h2` 패키지 분리
