--- src/main/java/persistence/sql/ddl/H2DropQueryBuilder.java
@@ -1,4 +1,0 @@
-package persistence.sql.ddl;
-
-public class H2DropQueryBuilder extends DropQueryBuilder {
-}
