--- src/main/java/persistence/sql/ddl/H2InsertQueryBuilder.java
@@ -1,7 +1,0 @@
-package persistence.sql.ddl;
-
-public class H2InsertQueryBuilder extends InsertQueryBuilder {
-    protected H2InsertQueryBuilder() {
-        super();
-    }
-}
