--- src/main/java/persistence/sql/ddl/H2SelectQueryBuilder.java
@@ -1,4 +1,0 @@
-package persistence.sql.ddl;
-
-public class H2SelectQueryBuilder extends SelectQueryBuilder {
-}
