+++ src/main/java/persistence/sql/ddl/h2/H2CreateDdlBuilder.java
@@ -1,6 +1,7 @@
-package persistence.sql.ddl;
+package persistence.sql.ddl.h2;
 
 import persistence.dialect.Database;
+import persistence.sql.ddl.CreateDdlBuilder;
 
 public class H2CreateDdlBuilder extends CreateDdlBuilder {
