+++ src/main/java/persistence/sql/ddl/h2/H2DeleteQueryBuilder.java
@@ -1,0 +1,6 @@
+package persistence.sql.ddl.h2;
+
+import persistence.sql.ddl.DeleteQueryBuilder;
+
+public class H2DeleteQueryBuilder extends DeleteQueryBuilder {
+}
