+++ src/main/java/persistence/sql/ddl/h2/H2DropQueryBuilder.java
@@ -1,0 +1,6 @@
+package persistence.sql.ddl.h2;
+
+import persistence.sql.ddl.DropQueryBuilder;
+
+public class H2DropQueryBuilder extends DropQueryBuilder {
+}
