+++ src/main/java/persistence/sql/ddl/h2/H2InsertQueryBuilder.java
@@ -1,0 +1,9 @@
+package persistence.sql.ddl.h2;
+
+import persistence.sql.ddl.InsertQueryBuilder;
+
+public class H2InsertQueryBuilder extends InsertQueryBuilder {
+    public H2InsertQueryBuilder() {
+        super();
+    }
+}
