+++ src/main/java/persistence/sql/ddl/h2/H2SelectQueryBuilder.java
@@ -1,0 +1,6 @@
+package persistence.sql.ddl.h2;
+
+import persistence.sql.ddl.SelectQueryBuilder;
+
+public class H2SelectQueryBuilder extends SelectQueryBuilder {
+}
