+++ src/test/java/persistence/dialect/DdlTypeTest.java
@@ -8,7 +8,7 @@ class DdlTypeTest {
 
     @Test
     void getTypeName() {
-        String pattern = "varchar(%l)";
+        String pattern = "varchar($l)";
         DdlType ddlType = new DdlType(pattern);
 
         Long size = 255L;
