+++ src/test/java/persistence/sql/ddl/DeleteQueryBuilderTest.java
@@ -1,6 +1,7 @@
 package persistence.sql.ddl;
 
 import org.junit.jupiter.api.Test;
+import persistence.sql.ddl.h2.H2DeleteQueryBuilder;
 
 import static org.assertj.core.api.Assertions.assertThat;
