+++ src/test/java/persistence/sql/ddl/InsertQueryBuilderTest.java
@@ -2,6 +2,7 @@ package persistence.sql.ddl;
 
 import domain.Person;
 import org.junit.jupiter.api.Test;
+import persistence.sql.ddl.h2.H2InsertQueryBuilder;
 
 import static org.assertj.core.api.Assertions.assertThat;
